Update operators to take floats.
git-svn-id: http://code.icecube.wisc.edu/svn/sandbox/jvansanten/MuonGun@101136 16731396-06f5-0310-8873-f7f720988828

+++ private/pybindings/Generator.cxx
@@ -20,9 +20,9 @@ void register_Generator()
 	    .def("generated_events", &GenerationProbability::GetGeneratedEvents)
 	    .add_property("total_events", &GenerationProbability::GetTotalEvents, &GenerationProbability::SetTotalEvents)
 	    .def("__add__",  (GenerationProbabilityPtr (*)(GenerationProbabilityPtr, GenerationProbabilityPtr))(&operator+))
-	    .def("__mul__",  (GenerationProbabilityPtr (*)(GenerationProbabilityPtr, size_t))(&operator*))
-	    .def("__rmul__", (GenerationProbabilityPtr (*)(GenerationProbabilityPtr, size_t))(&operator*))
-	    .def("__imul__", (GenerationProbabilityPtr (*)(GenerationProbabilityPtr, size_t))(&operator*=))
+	    .def("__mul__",  (GenerationProbabilityPtr (*)(GenerationProbabilityPtr, double))(&operator*))
+	    .def("__rmul__", (GenerationProbabilityPtr (*)(GenerationProbabilityPtr, double))(&operator*))
+	    .def("__imul__", (GenerationProbabilityPtr (*)(GenerationProbabilityPtr, double))(&operator*=))
 	;
 	implicitly_convertible<GenerationProbabilityPtr, GenerationProbabilityConstPtr>();
